Remove AppModule and replace with bootstrapApplication(AppComponent)
