+++ src/app/login/login-routing.module.ts
@@ -1,6 +1,5 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-
 import { LoginComponent } from './login.component';
 import { LoginHomeComponent } from './login-home/login-home.component';
